Seperated camera and search logic
