+++ app/src/main/res/layout/fragment_camera.xml
@@ -2,47 +2,14 @@
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-
+    android:id="@+id/cameraFragmentRl"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context=".fragments.CameraFragment">
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/cameraFragmentFl"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_alignBottom="@id/cameraFragmentFl"
-        android:layout_gravity="top">
-
-        <EditText
-            android:id="@+id/searchBarEt"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:autofillHints="@string/search_for_food"
-            android:hint="@string/search_for_food"
-            android:inputType="textAutoComplete"
-            android:minHeight="48dp"
-            android:selectAllOnFocus="true"
-            app:layout_constraintEnd_toStartOf="@id/submitSearchBtn"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <ImageButton
-            android:id="@+id/submitSearchBtn"
-            android:layout_width="40dp"
-            android:layout_height="40dp"
-            android:layout_gravity="start"
-            android:background="@drawable/camera"
-            android:enabled="true"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toEndOf="@id/searchBarEt"
-            app:layout_constraintTop_toTopOf="parent" />
-    </androidx.constraintlayout.widget.ConstraintLayout>
     <androidx.camera.view.PreviewView
         android:id="@+id/previewView"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_below="@+id/cameraFragmentFl"/>
+        android:layout_height="match_parent" />
 
     <TextView
         android:id="@+id/text_view"
